+++ inc-options/global_options.php
@@ -13,17 +13,17 @@ if ( ! function_exists( 'add_action' ) ) {
 		<div id="poststuff" class="ui-sortable meta-box-sortables">
 			<div class="postbox">
 				<div class="handlediv" title="<?php _e('Click to toggle'); ?>"><br/></div>
-				<h3 class="hndle" id="global_options"><?php _e('Global options', FB_ADMINIMIZE_TEXTDOMAIN ); ?></h3>
+				<h3 class="hndle" id="global_options"><?php _e('Global options', 'adminimize' ); ?></h3>
 				<div class="inside">
 					<br class="clear" />
 
 					<table summary="config_edit_post" class="widefat">
 						<thead>
 							<tr>
-								<th><?php _e('Option', FB_ADMINIMIZE_TEXTDOMAIN ); ?></th>
+								<th><?php _e('Option', 'adminimize' ); ?></th>
 								<?php
 									foreach ($user_roles_names as $role_name) { ?>
-										<th><?php _e('Deactivate for', FB_ADMINIMIZE_TEXTDOMAIN ); echo '<br/>' . $role_name; ?></th>
+										<th><?php _e('Deactivate for', 'adminimize' ); echo '<br/>' . $role_name; ?></th>
 								<?php } ?>
 							</tr>
 						</thead>
@@ -45,12 +45,12 @@ if ( ! function_exists( 'add_action' ) ) {
 							);
 							
 							$global_options_names = array(
-								esc_attr__('Admin Bar', FB_ADMINIMIZE_TEXTDOMAIN),
-								esc_attr__('Favorite Actions', FB_ADMINIMIZE_TEXTDOMAIN),
-								esc_attr__('Screen-Meta', FB_ADMINIMIZE_TEXTDOMAIN),
-								esc_attr__('Screen Options', FB_ADMINIMIZE_TEXTDOMAIN),
-								esc_attr__('Contextual Help', FB_ADMINIMIZE_TEXTDOMAIN),								esc_attr__('Admin Color Scheme', FB_ADMINIMIZE_TEXTDOMAIN),
-								esc_attr__( 'Admin Notices', FB_ADMINIMIZE_TEXTDOMAIN ),
+								esc_attr__('Admin Bar', 'adminimize'),
+								esc_attr__('Favorite Actions', 'adminimize'),
+								esc_attr__('Screen-Meta', 'adminimize'),
+								esc_attr__('Screen Options', 'adminimize'),
+								esc_attr__('Contextual Help', 'adminimize'),								esc_attr__('Admin Color Scheme', 'adminimize'),
+								esc_attr__( 'Admin Notices', 'adminimize' ),
 							);
 							
 							$_mw_adminimize_own_values = _mw_adminimize_get_option_value('_mw_adminimize_own_values');
@@ -102,25 +102,25 @@ if ( ! function_exists( 'add_action' ) ) {
 					<table summary="config_edit_post" class="widefat">
 						<thead>
 							<tr>
-								<th><?php _e('Your own options', FB_ADMINIMIZE_TEXTDOMAIN ); echo '<br />'; _e('ID or class', FB_ADMINIMIZE_TEXTDOMAIN ); ?></th>
-								<th><?php echo '<br />'; _e('Option', FB_ADMINIMIZE_TEXTDOMAIN ); ?></th>
+								<th><?php _e('Your own options', 'adminimize' ); echo '<br />'; _e('ID or class', 'adminimize' ); ?></th>
+								<th><?php echo '<br />'; _e('Option', 'adminimize' ); ?></th>
 							</tr>
 						</thead>
 
 						<tbody>
 							<tr valign="top">
-								<td colspan="2"><?php _e('It is possible to add your own IDs or classes from elements and tags. You can find IDs and classes with the FireBug Add-on for Firefox. Assign a value and the associate name per line.', FB_ADMINIMIZE_TEXTDOMAIN ); ?></td>
+								<td colspan="2"><?php _e('It is possible to add your own IDs or classes from elements and tags. You can find IDs and classes with the FireBug Add-on for Firefox. Assign a value and the associate name per line.', 'adminimize' ); ?></td>
 							</tr>
 							<tr valign="top">
 								<td>
 									<textarea name="_mw_adminimize_own_options" cols="60" rows="3" id="_mw_adminimize_own_options" style="width: 95%;" ><?php echo _mw_adminimize_get_option_value('_mw_adminimize_own_options'); ?></textarea>
 									<br />
-									<?php _e('Possible nomination for ID or class. Separate multiple nominations through a carriage return.', FB_ADMINIMIZE_TEXTDOMAIN ); ?>
+									<?php _e('Possible nomination for ID or class. Separate multiple nominations through a carriage return.', 'adminimize' ); ?>
 								</td>
 								<td>
 									<textarea class="code" name="_mw_adminimize_own_values" cols="60" rows="3" id="_mw_adminimize_own_values" style="width: 95%;" ><?php echo _mw_adminimize_get_option_value('_mw_adminimize_own_values'); ?></textarea>
 									<br />
-									<?php _e('Possible IDs or classes. Separate multiple values through a carriage return.', FB_ADMINIMIZE_TEXTDOMAIN ); ?>
+									<?php _e('Possible IDs or classes. Separate multiple values through a carriage return.', 'adminimize' ); ?>
 								</td>
 							</tr>
 						</tbody>
@@ -128,9 +128,9 @@ if ( ! function_exists( 'add_action' ) ) {
 					
 					<p id="submitbutton">
 						<input type="hidden" name="_mw_adminimize_action" value="_mw_adminimize_insert" />
-						<input class="button button-primary" type="submit" name="_mw_adminimize_save" value="<?php _e('Update Options', FB_ADMINIMIZE_TEXTDOMAIN ); ?> &raquo;" /><input type="hidden" name="page_options" value="'dofollow_timeout'" />
+						<input class="button button-primary" type="submit" name="_mw_adminimize_save" value="<?php _e('Update Options', 'adminimize' ); ?> &raquo;" /><input type="hidden" name="page_options" value="'dofollow_timeout'" />
 					</p>
-					<p><a class="alignright button" href="javascript:void(0);" onclick="window.scrollTo(0,0);" style="margin:3px 0 0 30px;"><?php _e('scroll to top', FB_ADMINIMIZE_TEXTDOMAIN); ?></a><br class="clear" /></p>
+					<p><a class="alignright button" href="javascript:void(0);" onclick="window.scrollTo(0,0);" style="margin:3px 0 0 30px;"><?php _e('scroll to top', 'adminimize'); ?></a><br class="clear" /></p>
 
 				</div>
 			</div>