+++ inc-options/theme_options.php
@@ -12,7 +12,7 @@ if ( ! function_exists( 'add_action' ) ) {
 <div id="poststuff" class="ui-sortable meta-box-sortables">
 	<div class="postbox">
 		<div class="handlediv" title="<?php _e( 'Click to toggle' ); ?>"><br /></div>
-		<h3 class="hndle" id="set_theme"><?php _e( 'Set Theme', FB_ADMINIMIZE_TEXTDOMAIN ) ?></h3>
+		<h3 class="hndle" id="set_theme"><?php _e( 'Set Theme', 'adminimize' ) ?></h3>
 
 		<div class="inside">
 			<br class="clear" />
@@ -24,13 +24,13 @@ if ( ! function_exists( 'add_action' ) ) {
 					<?php wp_nonce_field( 'mw_adminimize_nonce' ); ?>
 					<p><?php _e(
 							'For better peformance with many users on your blog; load only userlist, when you will change the theme options for users.',
-							FB_ADMINIMIZE_TEXTDOMAIN
+							'adminimize'
 						); ?></p>
 
 					<p id="submitbutton">
 						<input type="hidden" name="_mw_adminimize_action" value="_mw_adminimize_load_theme" />
 						<input type="submit" name="_mw_adminimize_load" value="<?php _e(
-							'Load User Data', FB_ADMINIMIZE_TEXTDOMAIN
+							'Load User Data', 'adminimize'
 						); ?> &raquo;" class="button button-primary" />
 					</p>
 				</form>
@@ -112,7 +112,7 @@ if ( ! function_exists( 'add_action' ) ) {
 						<input type="hidden" name="_mw_adminimize_action" value="_mw_adminimize_set_theme" />
 						<input type="hidden" name="_mw_adminimize_load" value="_mw_adminimize_load_theme" />
 						<input type="submit" name="_mw_adminimize_save" value="<?php _e(
-							'Set Theme', FB_ADMINIMIZE_TEXTDOMAIN
+							'Set Theme', 'adminimize'
 						); ?> &raquo;" class="button button-primary" />
 					</p>
 				</form>
@@ -120,7 +120,7 @@ if ( ! function_exists( 'add_action' ) ) {
 
 			<p>
 				<a class="alignright button" href="javascript:void(0);" onclick="window.scrollTo(0,0);" style="margin:3px 0 0 30px;"><?php _e(
-						'scroll to top', FB_ADMINIMIZE_TEXTDOMAIN
+						'scroll to top', 'adminimize'
 					); ?></a><br class="clear" /></p>
 		</div>
 	</div>